+++ src/Application/MainController.php
@@ -37,11 +37,10 @@ class MainController extends ControllerAbstract
 	{
 		$config = $this->get('config');
 		$template = $this->get('template');
-		$toolkit = $this->get('toolkit');
 
 		echo $template->render('index.php', [
 			'app_version' => $this->getAppVersion(),
-			'showBrowserExtensions' => ($toolkit->is_chrome() and $config->get('showBrowserExtensions')),
+			'showBrowserExtensions' => ($this->isUseragentChrome($_SERVER['HTTP_USER_AGENT']) and $config->get('showBrowserExtensions')),
 		]);
 	}
 }